pyproject.toml
[project]
name = "agentihooks"
version = "0.1.0"
description = "Hook system and MCP tool server for Claude Code agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "The Cloud Clock Work"},
]
keywords = ["claude-code", "mcp", "hooks", "agents", "agenticore"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.9"]
all = [
    "psycopg2-binary>=2.9.9",
    "pyyaml>=6.0",
    "redis>=5.0.0",
]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[tool.setuptools.packages.find]
include = ["hooks*"]

[project.urls]
Homepage = "https://github.com/The-Cloud-Clock-Work/agentihooks"
Repository = "https://github.com/The-Cloud-Clock-Work/agentihooks"
Issues = "https://github.com/The-Cloud-Clock-Work/agentihooks/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["unit: Unit tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = [
    "E501",   # line too long
    "F822",   # lazy-loaded names in __all__ (intentional for circular import avoidance)
    "E402",   # module-level imports not at top (conditional imports by design)
    "E741",   # ambiguous variable names (l, O, I) used in compact code
]

[tool.ruff.format]
quote-style = "double"
